Add initial socket auth
I've added the initial socket auth, when we connect to the server from the client,
the server will emit an auth request. The client responds to this by performing a
session confirmation request. If this is successful, the client will emit a successful
auth request along with the username and id (to be added).